+++ epd-ui/Module/childline.h
@@ -30,7 +30,7 @@ public:
                      QList<QString> stat_name_list,
                      QString timeS, QString timeE, QString price);
     void update_info(QString timeS, QString timeE, QString price);
-    void create_cheico(QList<int> che_index);
+    void create_cheico(QList<qint8> che_index);
     void updateTime(QString);
     void line_list_clear();
     void che_ico_list_clear();
